Refactored `misc.c/h` into - `math.c/h` -- contains mathematical functions - `Dev/simple_math.h` -- for example min/max, sign, ... - `Dev/misc.h` -- contains misc utility
