+++ Tests/test_dynstr.h
@@ -3,8 +3,8 @@
 //
 
 #include "../src/Dev/assert_that.h"
+#include "../src/Dev/misc.h"
 #include "../src/Dev/unit_tests.h"
-#include "../src/misc.h"
 #include "../src/string_utils.h"
 #include "../src/Structs/dynstring.h"
